+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 
 ### New
 - Added type checking of response object in `respond` functions, throwing an Error if it is not a valid `express` response object
+- Added `validateRequest()` to not only check for required query values but check the type of values
 
 ### Updated
 - Renamed `make` series of functions to `respond` to better reflect action of function. Legacy `make` functions are
   still available for backward compatibility but are undocumented
+- Added support for *JSON Server* style paging parameters (`_page` and `_limit`) in `calcPaging()`
 
 ## 1.2.3
 > Released 9 Jun 2021